\name{RFLPref}
\alias{RFLPref}
\docType{data}
\title{ Example data set for RFLP reference}
\description{
  This is an example data set for RFLP reference.
}
\usage{data(RFLPdata)}
\format{
  A data frame with 35 observations on the following five variables
  \describe{
    \item{\code{Sample}}{character: sample identifier. }
    \item{\code{Band}}{integer: band number. }
    \item{\code{MW}}{integer: molecular weight. }
    \item{\code{Taxonname}}{character: taxon name. }
    \item{\code{Accession}}{character: accession number. }
  }
}
\details{
  This example data set for RFLP reference consists of seven RFLP reference 
  samples. Taxon names are assigned by sequence comparison with GenBank 
  database (\url{http://www.ncbi.nlm.nih.gov/BLAST/}), and supplemented with 
  imaginary accession numbers.
}
\source{
  The data set was generated by F. Flessa.
}
%\references{ ??? }
\examples{
data(RFLPref)
str(RFLPref)
}
\keyword{datasets}
